fix: require passkey user verification
